Created header file drawgalaxy.h
Moved some additional stuff from drawgalaxy.cpp to navigation_system.cpp.
